refactor: move getValidatorsProfiles to base node
